requirements.txt
pandas>=1.2.5
google-analytics-data
